@@ -8,9 +8,6 @@ bootstrap = f"""#!/usr/bin/env zsh
 ## Description: Call script to start {__info__["name"]}.
 ## ******************************************************************************
 
-## Export PATH local user
-export PATH="$PATH:$HOME/.local/bin"
-
 function preexec() {{
   timer=$(date +%s)
 }}
@@ -19,11 +16,6 @@ function precmd() {{
     if [ $timer ]; then
       now=$(date +%s)
       elapsed=$(( $now - $timer ))
-      if [[ $elapsed -ge 60 ]]; then
-        took=$(date -d@$elapsed -u "+%Mm %Ss")
-      else
-        took=$(date -d@$elapsed -u "+%Ss")
-      fi
       unset timer
     fi
 }}
@@ -32,7 +24,7 @@ function precmd() {{
 function zshpower_precmd() {{
   state=$(which zshpower-draw > /dev/null 2>&1)
   if [ ! $? -ne 0 ]; then
-    PROMPT="$(zshpower-draw prompt $took)"
+    PROMPT="$(zshpower-draw prompt $elapsed)"
     RPROMPT="$(zshpower-draw rprompt)"
   else
     PROMPT='%F{{green}}%n%f@%F{{magenta}}%m%f %F{{blue}}%B%~%b%f %# '
