@@ -1,4 +1,4 @@
-from os import environ, geteuid
+from os import W_OK, access, environ, geteuid
 from pathlib import Path
 from subprocess import run
 
@@ -35,6 +35,9 @@ class Directory:
         self.hostname_enable = get_key(config, "hostname", "enable")
         self.truncate_value = get_key(config, "directory", "truncation_length")
         self.symbol = symbol_ssh(get_key(config, "directory", "symbol"), "")
+        self.lock_symbol = symbol_ssh(
+            get_key(config, "directory", "lock", "symbol"), ""
+        )
         self.color = (
             get_key(config, "directory", "color")
             if get_key(config, "general", "color", "enable") is True
@@ -71,7 +74,13 @@ class Directory:
 
         if dir_truncate.split("/")[-1:] == str(Path.home()).split("/")[-1:]:
             dir_truncate = "~"
+
+        if not access(get_pwd(), W_OK):
+            lock_symbol = f"{Color('red')}{self.lock_symbol}{Color().NONE}"
+        else:
+            lock_symbol = ""
         return (
             f"{prefix}{Color(self.color)}{self.symbol}"
             f"{dir_truncate}{space_elem}{Color().NONE}"
+            f"{lock_symbol}"
         )
