@@ -1,11 +1,11 @@
-from typing import Any
+from datetime import timedelta
 
 from snakypy.zshpower.prompt.sections.utils import Color, symbol_ssh
 from snakypy.zshpower.utils.catch import get_key
 
 
 class Took:
-    def __init__(self, config: dict, took: Any = 0):
+    def __init__(self, config: dict, elapsed: float = 0.0):
         self.enable = get_key(config, "took", "enable")
         self.symbol: str = symbol_ssh(get_key(config, "took", "symbol"), "")
         self.text: str = get_key(config, "took", "text")
@@ -16,22 +16,19 @@ class Took:
         )
         self.involved: str = get_key(config, "took", "involved")
         self.show_greater_than: str = get_key(config, "took", "show_greater_than")
-        self.took: str = str(took)
+        self.elapsed: float = elapsed
 
     def format_took(self) -> str:
-        if self.took and len(self.took) == 7:
-            if self.took[0] == "0" and self.took[4] == "0":
-                return f"{self.took[1:3]} {self.took[5:]}"
-            elif self.took[0] == "0":
-                return f"{self.took[1:3]} {self.took[4:]}"
-            elif self.took[4] == "0":
-                return f"{self.took[0:3]} {self.took[5:]}"
+        # split_timer out: ["00", "00", "00"] (H:M:S)
+        split_timer = str(timedelta(seconds=self.elapsed)).split(":")
+        if int(split_timer[0]) > 0:
+            return f"{split_timer[0]}h {split_timer[1]}m {split_timer[2]}s"
+        elif int(split_timer[1]) > 0:
+            return f"{split_timer[1]}m {split_timer[2]}s"
         else:
-            if self.took[0] == "0":
-                return self.took[1:3]
-        return self.took
+            return f"{split_timer[2]}s"
 
-    def show(self):
+    def show(self) -> str:
         took_current = (
             f" {Color(self.color)}{self.symbol}{self.text} "
             f"{Color().NONE}{self.format_took()}"
@@ -45,11 +42,11 @@ class Took:
 
     def __str__(self):
         if self.enable:
-            if len(str(self.took)) == 7:
+            split_timer = self.format_took().split(" ")
+            seconds = int(self.format_took().split(" ")[-1].replace("s", ""))
+            if len(split_timer) == 3 or len(split_timer) == 2:
                 return self.show()
             else:
-                verify_seconds = f"{self.format_took()[:-1]}"
-                if verify_seconds > str(self.show_greater_than):
+                if seconds > float(self.show_greater_than):
                     return self.show()
-
         return ""
